requirements.txt
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1743190139881/work
annotated-types @ file:///work/perseverance-python-buildout/croot/annotated-types_1728386958747/work
archspec @ file:///croot/archspec_1709217642129/work
blinker==1.9.0
boltons @ file:///croot/boltons_1737061692168/work
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///croot/certifi_1745939216646/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.2.1
conda @ file:///croot/conda_1743715288469/work/conda-src
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1744662537425/work
conda-anaconda-tos @ file:///croot/conda-anaconda-tos_1744823864153/work
conda-content-trust @ file:///work/perseverance-python-buildout/croot/conda-content-trust_1728487058234/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1745607008911/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda_package_streaming @ file:///croot/conda-package-streaming_1731366181659/work
cryptography @ file:///croot/cryptography_1740577825284/work
dash==3.0.4
distro @ file:///work/perseverance-python-buildout/croot/distro_1728396110052/work
Flask==3.0.3
frozendict @ file:///work/perseverance-python-buildout/croot/frozendict_1728497542215/work
gunicorn==23.0.0
idna @ file:///work/perseverance-python-buildout/croot/idna_1728385935861/work
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch @ file:///work/perseverance-python-buildout/croot/jsonpatch_1728399595941/work
jsonpointer==2.1
libmambapy @ file:///croot/mamba-split_1734469461757/work/libmambapy
markdown-it-py @ file:///work/perseverance-python-buildout/croot/markdown-it-py_1728387994477/work
MarkupSafe==3.0.2
mdurl @ file:///work/perseverance-python-buildout/croot/mdurl_1728387286143/work
menuinst @ file:///croot/menuinst_1738943416351/work
narwhals==1.44.0
nest-asyncio==1.6.0
numpy==2.3.1
packaging @ file:///croot/packaging_1734472117206/work
pandas==2.3.0
platformdirs @ file:///croot/platformdirs_1744273042065/work
plotly==6.1.2
pluggy @ file:///croot/pluggy_1733169602837/work
pycosat @ file:///croot/pycosat_1736868416091/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic @ file:///croot/pydantic_1734736067156/work
pydantic_core @ file:///croot/pydantic-core_1734726052986/work
Pygments @ file:///croot/pygments_1744664109463/work
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1728386193338/work
python-dateutil==2.9.0.post0
pytz==2025.2
requests @ file:///croot/requests_1730999120400/work
retrying==1.4.0
rich @ file:///croot/rich_1732638981168/work
ruamel.yaml @ file:///croot/ruamel.yaml_1745960305322/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1745937152469/work
setuptools==78.1.1
six==1.17.0
tqdm @ file:///croot/tqdm_1738943501192/work
truststore @ file:///croot/truststore_1736550121485/work
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.2
urllib3 @ file:///croot/urllib3_1737133630106/work
Werkzeug==3.0.6
wheel==0.45.1
zipp==3.23.0
zstandard @ file:///croot/zstandard_1731356346222/work
